requirement.txt
django 
djangorestframework 
djangorestframework-simplejwt 
opencv-python 
numpy 
deepface 
pillow
tf-keras
celery 
redis
flower